Finished refactoring users
