+++ routes/users.js
@@ -5,31 +5,32 @@ module.exports = function (app, router, db, models) {
   router.route('/users')
     .get(function(req, res) {
       //return all users
-      try {
-        models.sequelize.sync(/*{force:true}*/).then(() => {
-          db.getUsers()
-            .then(aa => res.json({ status: true, body: aa }))  // DOC: what does aa stand for? Is it a string or an Object? (json {})
-        });
-      } catch (e) {
-        console.log(e);
-        res.json({ status: false, body: {error: "message"}});
-      }
+      db.getUsers()
+        .then(users => res.json({ status: true, body: users }))
+  //     try {
+  //       models.sequelize.sync(/*{force:true}*/).then(() => {
+  // // DOC: what does aa stand for? Is it a string or an Object? (json {})
+  //       });
+  //     } catch (e) {
+  //       console.log(e);
+  //       res.json({ status: false, body: {error: "message"}});
+  //     }
     })
 
 
   router.route('/users/:user_type/:user_id')
     .get(function(req, res) {
       //return user with req.params.user_type and req.params.user_id *** not sure if needed
-
-      try {
-        models.sequelize.sync(/*{force:true}*/).then(() => {
-          db.getUserByTypeAndId(req.params.user_type, req.params.user_id)
-            .then(aa => res.json({ status: true, body: aa }))
-        });
-      } catch (e) {
-        console.log(e);
-        res.json({ status: false, body: {error: "message"}});
-      }
+      db.getUserById(req.params.user_id)
+        .then(user => res.json({ status: true, body: user }))
+      // try {
+      //   models.sequelize.sync(/*{force:true}*/).then(() => {
+      //
+      //   });
+      // } catch (e) {
+      //   console.log(e);
+      //   res.json({ status: false, body: {error: "message"}});
+      // }
     })
     .post(function(req, res) {
       //add user to DB with req.params.user_type and req.params.user_id
